+++ lib/features/home/service/components/profile_details.dart
@@ -21,24 +21,19 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <https://www.gnu.org/licenses/>.
 ///
-/// Authors: Ashley Tang
+/// Authors: Ashley Tang, Tony Chen
 
 library;
 
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
 
-import 'package:markdown_tooltip/markdown_tooltip.dart';
 import 'package:solidpod/solidpod.dart';
 
-import 'package:healthpod/constants/appointment.dart';
-import 'package:healthpod/constants/paths.dart';
+import 'package:healthpod/features/home/service/components/profile/profile_data_manager.dart';
+import 'package:healthpod/features/home/service/components/profile/profile_edit_dialog.dart';
+import 'package:healthpod/features/home/service/components/profile/profile_photo_manager.dart';
+import 'package:healthpod/features/home/service/components/profile/profile_ui_components.dart';
 import 'package:healthpod/theme/card_style.dart';
-import 'package:healthpod/utils/fetch_profile_data.dart';
-import 'package:healthpod/utils/format_timestamp_for_filename.dart';
-import 'package:healthpod/utils/is_logged_in.dart';
-import 'package:healthpod/utils/profile_photo_handler.dart';
 
 /// A widget that combines user avatar and name with personal identification details.
 /// This integrated component displays all user profile information in a single card.
@@ -78,22 +73,13 @@ class ProfileDetails extends StatefulWidget {
 class _ProfileDetailsState extends State<ProfileDetails> {
   // Controllers for the editable fields.
 
-  late TextEditingController _nameController;
-  late TextEditingController _addressController;
-  late TextEditingController _bestContactPhoneController;
-  late TextEditingController _bestContactEmailController;
-  late TextEditingController _emergencyNameController;
-  late TextEditingController _emergencyPhoneController;
-  late TextEditingController _alternativeContactNumberController;
-  late TextEditingController _emailController;
-  late TextEditingController _dateOfBirthController;
-  late TextEditingController _genderController;
+  late final Map<TextEditingController, String> _controllers;
 
   final _formKey = GlobalKey<FormState>();
   bool _isLoading = true;
   bool _isSaving = false;
   bool _isLoadingPhoto = true;
-  bool _isUploadingPhoto = false;
+  final bool _isUploadingPhoto = false;
   ImageProvider? _profilePhoto;
 
   // Holds full profile data.
@@ -111,16 +97,18 @@ class _ProfileDetailsState extends State<ProfileDetails> {
   /// Initialise all text controllers.
 
   void _initializeControllers() {
-    _nameController = TextEditingController();
-    _addressController = TextEditingController();
-    _bestContactPhoneController = TextEditingController();
-    _bestContactEmailController = TextEditingController();
-    _emergencyNameController = TextEditingController();
-    _emergencyPhoneController = TextEditingController();
-    _alternativeContactNumberController = TextEditingController();
-    _emailController = TextEditingController();
-    _dateOfBirthController = TextEditingController();
-    _genderController = TextEditingController();
+    _controllers = {
+      TextEditingController(): 'name',
+      TextEditingController(): 'address',
+      TextEditingController(): 'bestContactPhone',
+      TextEditingController(): 'bestContactEmail',
+      TextEditingController(): 'emergencyName',
+      TextEditingController(): 'emergencyPhone',
+      TextEditingController(): 'alternativeContactNumber',
+      TextEditingController(): 'email',
+      TextEditingController(): 'dateOfBirth',
+      TextEditingController(): 'gender',
+    };
   }
 
   /// Load profile data from the pod and update state.
@@ -131,52 +119,21 @@ class _ProfileDetailsState extends State<ProfileDetails> {
     });
 
     try {
-      // First check if user is logged in.
+      final profileData = await ProfileDataManager.loadProfileData(context);
+      _profileData = profileData;
 
-      final loggedIn = await isLoggedIn();
+      setState(() {
+        // Populate controllers with profile data or defaults.
 
-      if (loggedIn) {
-        if (!mounted) return;
-        // Fetch profile data using utility function.
-
-        final profileData = await fetchProfileData(context);
-        _profileData = profileData;
-
-        setState(() {
-          // Populate controllers with profile data or defaults.
-
-          _nameController.text = profileData['name'] ?? userName;
-          _addressController.text = profileData['address'] ?? '';
-          _bestContactPhoneController.text =
-              profileData['bestContactPhone'] ?? '';
-          _bestContactEmailController.text =
-              profileData['bestContactEmail'] ?? '';
-          _emergencyNameController.text = profileData['emergencyName'] ?? '';
-          _emergencyPhoneController.text = profileData['emergencyPhone'] ?? '';
-          _alternativeContactNumberController.text =
-              profileData['alternativeContactNumber'] ?? '';
-          _emailController.text = profileData['email'] ?? '';
-          _dateOfBirthController.text = profileData['dateOfBirth'] ?? '';
-          _genderController.text = profileData['gender'] ?? '';
-          _isLoading = false;
+        _controllers.forEach((controller, fieldName) {
+          String value = profileData[fieldName] ?? '';
+          if (fieldName == 'name' && value.isEmpty) {
+            value = ProfileDataManager.getDefaultName();
+          }
+          controller.text = value;
         });
-      } else {
-        // User not logged in, just use default values.
-
-        setState(() {
-          _nameController.text = userName;
-          _addressController.text = '';
-          _bestContactPhoneController.text = '';
-          _bestContactEmailController.text = '';
-          _emergencyNameController.text = '';
-          _emergencyPhoneController.text = '';
-          _alternativeContactNumberController.text = '';
-          _emailController.text = '';
-          _dateOfBirthController.text = '';
-          _genderController.text = '';
-          _isLoading = false;
-        });
-      }
+        _isLoading = false;
+      });
     } catch (e) {
       setState(() {
         _isLoading = false;
@@ -192,7 +149,7 @@ class _ProfileDetailsState extends State<ProfileDetails> {
     });
 
     try {
-      final photoProvider = await ProfilePhotoHandler.getProfilePhoto(context);
+      final photoProvider = await ProfilePhotoManager.loadProfilePhoto(context);
       setState(() {
         _profilePhoto = photoProvider;
         _isLoadingPhoto = false;
@@ -208,9 +165,15 @@ class _ProfileDetailsState extends State<ProfileDetails> {
   /// Save profile data to the pod.
 
   Future<void> _saveProfileData() async {
+    // Get the name controller.
+
+    final nameController = _controllers.keys.firstWhere(
+      (controller) => _controllers[controller] == 'name',
+    );
+
     // Validate only the name field.
 
-    if (_nameController.text.trim().isEmpty) {
+    if (nameController.text.trim().isEmpty) {
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           const SnackBar(content: Text('Name is required')),
@@ -221,7 +184,7 @@ class _ProfileDetailsState extends State<ProfileDetails> {
 
     // Skip if no changes detected.
 
-    if (!_hasDataChanged()) {
+    if (!ProfileDataManager.hasDataChanged(_controllers, _profileData)) {
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           const SnackBar(content: Text('No changes detected')),
@@ -237,23 +200,17 @@ class _ProfileDetailsState extends State<ProfileDetails> {
     try {
       // Prepare data for saving.
 
-      final updatedData = {
-        'name': _nameController.text.trim(),
-        'address': _addressController.text.trim(),
-        'bestContactPhone': _bestContactPhoneController.text.trim(),
-        'bestContactEmail': _bestContactEmailController.text.trim(),
-        'emergencyName': _emergencyNameController.text.trim(),
-        'emergencyPhone': _emergencyPhoneController.text.trim(),
-        'alternativeContactNumber':
-            _alternativeContactNumberController.text.trim(),
-        'email': _emailController.text.trim(),
-        'dateOfBirth': _dateOfBirthController.text.trim(),
-        'gender': _genderController.text.trim(),
-      };
+      final updatedData = <String, dynamic>{};
+      _controllers.forEach((controller, fieldName) {
+        updatedData[fieldName] = controller.text.trim();
+      });
 
-      // Try to update existing profile file first, create new one only if none exists.
+      // Save using ProfileDataManager.
 
-      final result = await _updateOrCreateProfileFile(updatedData);
+      final result = await ProfileDataManager.saveProfileData(
+        updatedData,
+        context,
+      );
 
       if (result != SolidFunctionCallStatus.success) {
         throw Exception('Failed to save profile data: $result');
@@ -307,130 +264,13 @@ class _ProfileDetailsState extends State<ProfileDetails> {
     }
   }
 
-  /// Updates existing profile file or creates a new one if none exists.
-
-  Future<SolidFunctionCallStatus> _updateOrCreateProfileFile(
-    Map<String, dynamic> updatedData,
-  ) async {
-    try {
-      // First, try to find an existing profile file.
-
-      final existingFile = await _findExistingProfileFile();
-
-      String filename;
-      if (existingFile != null) {
-        // Use the existing filename to update the file.
-
-        filename = existingFile;
-      } else {
-        // Create new filename only if no existing file found.
-
-        final timestamp = formatTimestampForFilename(DateTime.now());
-        filename = 'profile_$timestamp.json.enc.ttl';
-      }
-
-      // Create JSON data structure matching other successful implementations.
-      final profileData = {
-        'timestamp': DateTime.now().toIso8601String(),
-        'responses': updatedData,
-      };
-
-      // Check if context is still valid before using it.
-
-      if (!mounted) {
-        debugPrint('❌ Context no longer mounted during save');
-        return SolidFunctionCallStatus.fail;
-      }
-
-      // Use direct writePod call with relative path to match read operations.
-
-      final result = await writePod(
-        'profile/$filename',
-        json.encode(profileData),
-        context,
-        const Text('Saving profile data'),
-        encrypted: true,
-      );
-
-      return result;
-    } on Exception catch (e) {
-      debugPrint('Exception saving profile: $e');
-      return SolidFunctionCallStatus.fail;
-    } catch (e) {
-      debugPrint('Unexpected error saving profile: $e');
-      return SolidFunctionCallStatus.fail;
-    }
-  }
-
-  /// Finds an existing profile file to update, or returns null if none exists.
-
-  Future<String?> _findExistingProfileFile() async {
-    try {
-      // Get all files in the profile directory.
-
-      final dirUrl = await getDirUrl('$basePath/profile');
-      final resources = await getResourcesInContainer(dirUrl);
-
-      // Find all profile files with the expected extension.
-
-      final profileFiles = resources.files
-          .where(
-            (file) =>
-                file.startsWith('profile_') &&
-                !file.startsWith('profile_photo_') &&
-                file.endsWith('.json.enc.ttl'),
-          )
-          .toList();
-
-      if (profileFiles.isEmpty) {
-        return null;
-      }
-
-      // Return the most recent profile file (sorted by filename).
-
-      profileFiles.sort((a, b) => b.compareTo(a));
-      return profileFiles.first;
-    } catch (e) {
-      debugPrint('Error finding existing profile file: $e');
-      return null;
-    }
-  }
-
-  /// Check if any profile data has been changed compared to stored data.
-
-  bool _hasDataChanged() {
-    return _nameController.text.trim() != (_profileData['name'] ?? '') ||
-        _addressController.text.trim() != (_profileData['address'] ?? '') ||
-        _bestContactPhoneController.text.trim() !=
-            (_profileData['bestContactPhone'] ?? '') ||
-        _bestContactEmailController.text.trim() !=
-            (_profileData['bestContactEmail'] ?? '') ||
-        _emergencyNameController.text.trim() !=
-            (_profileData['emergencyName'] ?? '') ||
-        _emergencyPhoneController.text.trim() !=
-            (_profileData['emergencyPhone'] ?? '') ||
-        _alternativeContactNumberController.text.trim() !=
-            (_profileData['alternativeContactNumber'] ?? '') ||
-        _emailController.text.trim() != (_profileData['email'] ?? '') ||
-        _dateOfBirthController.text.trim() !=
-            (_profileData['dateOfBirth'] ?? '') ||
-        _genderController.text.trim() != (_profileData['gender'] ?? '');
-  }
-
   @override
   void dispose() {
     // Clean up all controllers to prevent memory leaks.
 
-    _nameController.dispose();
-    _addressController.dispose();
-    _bestContactPhoneController.dispose();
-    _bestContactEmailController.dispose();
-    _emergencyNameController.dispose();
-    _emergencyPhoneController.dispose();
-    _alternativeContactNumberController.dispose();
-    _emailController.dispose();
-    _dateOfBirthController.dispose();
-    _genderController.dispose();
+    for (var controller in _controllers.keys) {
+      controller.dispose();
+    }
     super.dispose();
   }
 
@@ -454,542 +294,37 @@ class _ProfileDetailsState extends State<ProfileDetails> {
   /// Show dialog for editing profile details.
 
   Future<void> _showEditDialog() async {
-    // Create temporary controllers for dialog fields.
+    // Prepare current data for the dialog.
 
-    final tempNameController =
-        TextEditingController(text: _nameController.text);
-    final tempAddressController =
-        TextEditingController(text: _addressController.text);
-    final tempBestContactPhoneController =
-        TextEditingController(text: _bestContactPhoneController.text);
-    final tempBestContactEmailController =
-        TextEditingController(text: _bestContactEmailController.text);
-    final tempEmergencyNameController =
-        TextEditingController(text: _emergencyNameController.text);
-    final tempEmergencyPhoneController =
-        TextEditingController(text: _emergencyPhoneController.text);
-    final tempAlternativeContactNumberController =
-        TextEditingController(text: _alternativeContactNumberController.text);
-    final tempEmailController =
-        TextEditingController(text: _emailController.text);
-    final tempDateOfBirthController =
-        TextEditingController(text: _dateOfBirthController.text);
-    final tempGenderController =
-        TextEditingController(text: _genderController.text);
+    final currentData = <String, String>{};
+    _controllers.forEach((controller, fieldName) {
+      currentData[fieldName] = controller.text;
+    });
 
-    final formKey = GlobalKey<FormState>();
+    // Show the edit dialog.
 
-    // Show dialog with edit form.
+    final updatedData = await ProfileEditDialog.show(context, currentData);
 
-    final result = await showDialog<bool>(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Edit Profile Details'),
-          content: SingleChildScrollView(
-            child: Form(
-              key: formKey,
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  const Text('Name'),
-                  TextFormField(
-                    controller: tempNameController,
-                    decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      contentPadding:
-                          EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                    ),
-                    validator: (value) {
-                      if (value == null || value.trim().isEmpty) {
-                        return 'Name is required';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Address'),
-                  TextFormField(
-                    controller: tempAddressController,
-                    decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      contentPadding:
-                          EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                    ),
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Phone'),
-                  MarkdownTooltip(
-                    message: '''
+    // If user saved changes, update controllers and save data.
 
-                    **Valid Phone Number Formats:**
-
-                    - **Australian Mobile:** +61 4XX XXX XXX or 04XX XXX XXX
-                    - **Australian Landline:** +61 X XXXX XXXX or 0X XXXX XXXX
-                    - **International:** +[country code] followed by number
-
-                    Spaces, dashes and parentheses are allowed.
-
-                    ''',
-                    child: TextFormField(
-                      controller: tempBestContactPhoneController,
-                      decoration: const InputDecoration(
-                        border: OutlineInputBorder(),
-                        contentPadding:
-                            EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                        hintText: 'e.g. +61 4 1234 5678 or 04 1234 5678',
-                        suffixIcon: Icon(Icons.info_outline),
-                      ),
-                      validator: _validatePhone,
-                      keyboardType: TextInputType.phone,
-                    ),
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Emergency Name'),
-                  TextFormField(
-                    controller: tempEmergencyNameController,
-                    decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      contentPadding:
-                          EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                    ),
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Emergency Phone'),
-                  MarkdownTooltip(
-                    message: '''
-
-                    **Valid Phone Number Formats:**
-
-                    - **Australian Mobile:** +61 4XX XXX XXX or 04XX XXX XXX
-                    - **Australian Landline:** +61 X XXXX XXXX or 0X XXXX XXXX
-                    - **International:** +[country code] followed by number
-
-                    Spaces, dashes and parentheses are allowed.
-
-                    ''',
-                    child: TextFormField(
-                      controller: tempEmergencyPhoneController,
-                      decoration: const InputDecoration(
-                        border: OutlineInputBorder(),
-                        contentPadding:
-                            EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                        hintText: 'e.g. +61 4 1234 5678 or 04 1234 5678',
-                        suffixIcon: Icon(Icons.info_outline),
-                      ),
-                      validator: _validatePhone,
-                      keyboardType: TextInputType.phone,
-                    ),
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Alternative Phone'),
-                  MarkdownTooltip(
-                    message: '''
-
-                    **Valid Phone Number Formats:**
-
-                    - **Australian Mobile:** +61 4XX XXX XXX or 04XX XXX XXX
-                    - **Australian Landline:** +61 X XXXX XXXX or 0X XXXX XXXX
-                    - **International:** +[country code] followed by number
-
-                    Spaces, dashes and parentheses are allowed.
-
-                    ''',
-                    child: TextFormField(
-                      controller: tempAlternativeContactNumberController,
-                      decoration: const InputDecoration(
-                        border: OutlineInputBorder(),
-                        contentPadding:
-                            EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                        hintText: 'e.g. +61 4 1234 5678 or 04 1234 5678',
-                        suffixIcon: Icon(Icons.info_outline),
-                      ),
-                      validator: _validatePhone,
-                      keyboardType: TextInputType.phone,
-                    ),
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Email'),
-                  TextFormField(
-                    controller: tempEmailController,
-                    decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      contentPadding:
-                          EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                    ),
-                    validator: _validateEmail,
-                    keyboardType: TextInputType.emailAddress,
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Date of Birth'),
-                  Row(
-                    children: [
-                      Expanded(
-                        child: InkWell(
-                          onTap: () async {
-                            // Show date picker for selecting date of birth.
-
-                            final DateTime? picked = await showDatePicker(
-                              context: context,
-                              initialDate: _parseDateOrDefault(
-                                tempDateOfBirthController.text,
-                              ),
-                              firstDate: DateTime(1900),
-                              lastDate: DateTime.now(),
-                            );
-                            if (picked != null) {
-                              tempDateOfBirthController.text =
-                                  _formatDate(picked);
-                            }
-                          },
-                          child: AbsorbPointer(
-                            child: TextFormField(
-                              controller: tempDateOfBirthController,
-                              decoration: const InputDecoration(
-                                border: OutlineInputBorder(),
-                                contentPadding: EdgeInsets.symmetric(
-                                  horizontal: 8,
-                                  vertical: 4,
-                                ),
-                                suffixIcon: Icon(Icons.calendar_today),
-                              ),
-                              keyboardType: TextInputType.datetime,
-                            ),
-                          ),
-                        ),
-                      ),
-                      if (tempDateOfBirthController.text.isNotEmpty)
-                        IconButton(
-                          icon: const Icon(Icons.clear),
-                          onPressed: () {
-                            tempDateOfBirthController.clear();
-                          },
-                        ),
-                    ],
-                  ),
-                  const SizedBox(height: 12),
-                  const Text('Gender'),
-                  Row(
-                    children: [
-                      Expanded(
-                        child: DropdownButtonFormField<String?>(
-                          initialValue: tempGenderController.text.isEmpty
-                              ? null
-                              : tempGenderController.text,
-                          decoration: const InputDecoration(
-                            border: OutlineInputBorder(),
-                            contentPadding: EdgeInsets.symmetric(
-                              horizontal: 8,
-                              vertical: 4,
-                            ),
-                          ),
-                          items: const [
-                            DropdownMenuItem(
-                              value: null,
-                              child: Text('Select gender'),
-                            ),
-                            DropdownMenuItem(
-                              value: 'Male',
-                              child: Text('Male'),
-                            ),
-                            DropdownMenuItem(
-                              value: 'Female',
-                              child: Text('Female'),
-                            ),
-                            DropdownMenuItem(
-                              value: 'Non-binary',
-                              child: Text('Non-binary'),
-                            ),
-                            DropdownMenuItem(
-                              value: 'Prefer not to say',
-                              child: Text('Prefer not to say'),
-                            ),
-                          ],
-                          onChanged: (value) {
-                            tempGenderController.text = value ?? '';
-                          },
-                        ),
-                      ),
-                      if (tempGenderController.text.isNotEmpty)
-                        IconButton(
-                          icon: const Icon(Icons.clear),
-                          onPressed: () {
-                            tempGenderController.clear();
-                          },
-                        ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(false),
-              child: const Text('Cancel'),
-            ),
-            TextButton(
-              onPressed: () {
-                if (formKey.currentState!.validate()) {
-                  Navigator.of(context).pop(true);
-                }
-              },
-              child: const Text('Save'),
-            ),
-          ],
-        );
-      },
-    );
-
-    // If user confirmed, update main controllers and save data.
-
-    if (result == true) {
+    if (updatedData != null) {
       setState(() {
-        _nameController.text = tempNameController.text;
-        _addressController.text = tempAddressController.text;
-        _bestContactPhoneController.text = tempBestContactPhoneController.text;
-        _bestContactEmailController.text = tempBestContactEmailController.text;
-        _emergencyNameController.text = tempEmergencyNameController.text;
-        _emergencyPhoneController.text = tempEmergencyPhoneController.text;
-        _alternativeContactNumberController.text =
-            tempAlternativeContactNumberController.text;
-        _emailController.text = tempEmailController.text;
-        _dateOfBirthController.text = tempDateOfBirthController.text;
-        _genderController.text = tempGenderController.text;
+        _controllers.forEach((controller, fieldName) {
+          controller.text = updatedData[fieldName] ?? '';
+        });
       });
 
       await _saveProfileData();
-    }
-
-    // Clean up temporary controllers.
-
-    tempNameController.dispose();
-    tempAddressController.dispose();
-    tempBestContactPhoneController.dispose();
-    tempBestContactEmailController.dispose();
-    tempEmergencyNameController.dispose();
-    tempEmergencyPhoneController.dispose();
-    tempAlternativeContactNumberController.dispose();
-    tempEmailController.dispose();
-    tempDateOfBirthController.dispose();
-    tempGenderController.dispose();
-  }
-
-  /// Parse a date string into DateTime or return a default date.
-
-  DateTime _parseDateOrDefault(String dateStr) {
-    try {
-      final parts = dateStr.split('-');
-      if (parts.length == 3) {
-        return DateTime(
-          int.parse(parts[0]),
-          int.parse(parts[1]),
-          int.parse(parts[2]),
-        );
-      }
-    } catch (e) {
-      //debugPrint('Error parsing date: $e');
-    }
-    // Return a default date (30 years ago)
-    return DateTime.now().subtract(const Duration(days: 365 * 30));
-  }
-
-  /// Format a DateTime as YYYY-MM-DD.
-
-  String _formatDate(DateTime date) {
-    return '${date.year}-${date.month.toString().padLeft(2, '0')}-${date.day.toString().padLeft(2, '0')}';
-  }
-
-  /// Validate email format - optional field.
-
-  String? _validateEmail(String? value) {
-    if (value == null || value.isEmpty) return null;
-    final emailRegex = RegExp(r'^[^@]+@[^@]+\.[^@]+$');
-    if (!emailRegex.hasMatch(value)) return 'Enter a valid email address';
-    return null;
-  }
-
-  /// Validate phone number format - optional field.
-
-  String? _validatePhone(String? value) {
-    if (value == null || value.isEmpty) return null;
-
-    // Clean the input by removing spaces, dashes and parentheses.
-
-    final cleanedValue = value.replaceAll(RegExp(r'[\s\-()]'), '');
-    final australianPhoneRegex = RegExp(r'^(\+61|0)[0-9]{9,10}$');
-    final internationalPhoneRegex = RegExp(r'^\+[0-9]{10,14}$');
-
-    if (!australianPhoneRegex.hasMatch(cleanedValue) &&
-        !internationalPhoneRegex.hasMatch(cleanedValue)) {
-      return 'Enter a valid phone number (e.g. +61 4 1234 5678 or 04 1234 5678)';
-    }
-    return null;
-  }
-
-  /// Show dialog for selecting profile photo options.
-
-  Future<void> _showPhotoOptionsDialog() async {
-    if (_isLoading || _isSaving || _isLoadingPhoto || _isUploadingPhoto) {
-      return;
-    }
-
-    await showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Profile Photo'),
-          content: SingleChildScrollView(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                // Display current photo or avatar.
-
-                SizedBox(
-                  height: 100,
-                  width: 100,
-                  child: ProfilePhotoHandler.buildProfileAvatar(
-                    context: context,
-                    photo: _profilePhoto,
-                    name: _nameController.text,
-                    radius: 50,
-                  ),
-                ),
-                const SizedBox(height: 20),
-                // Photo action buttons.
-
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    ElevatedButton.icon(
-                      onPressed: () {
-                        Navigator.pop(context);
-                        _handlePhotoUpload();
-                      },
-                      icon: const Icon(Icons.photo_camera),
-                      label: const Text('Upload New'),
-                    ),
-                    if (_profilePhoto != null)
-                      ElevatedButton.icon(
-                        onPressed: () {
-                          Navigator.pop(context);
-                          _handlePhotoDelete();
-                        },
-                        icon: const Icon(Icons.delete),
-                        label: const Text('Remove'),
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Theme.of(context).colorScheme.error,
-                          foregroundColor:
-                              Theme.of(context).colorScheme.onError,
-                        ),
-                      ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(),
-              child: const Text('Close'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  /// Handle photo upload process.
-
-  Future<void> _handlePhotoUpload() async {
-    setState(() {
-      _isUploadingPhoto = true;
-    });
-
-    try {
-      final imageFile = await ProfilePhotoHandler.pickProfilePhoto();
-
-      if (imageFile != null && mounted) {
-        final success = await ProfilePhotoHandler.uploadProfilePhoto(
-          imageFile,
-          context,
-        );
-
-        if (success && mounted) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('Profile photo uploaded successfully'),
-              backgroundColor: Colors.green,
-            ),
-          );
-
-          // Cleanup old photos.
-
-          await ProfilePhotoHandler.cleanupOldProfilePhotos(context);
-
-          // Reload the photo.
-
-          await _loadProfilePhoto();
-
-          // Notify parent of data change.
-
-          widget.onDataChanged();
-        }
-      }
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isUploadingPhoto = false;
-        });
-      }
-    }
-  }
-
-  /// Handle photo deletion.
-
-  Future<void> _handlePhotoDelete() async {
-    setState(() {
-      _isUploadingPhoto = true;
-    });
-
-    try {
-      if (mounted) {
-        final success = await ProfilePhotoHandler.deleteProfilePhoto(context);
-
-        if (success && mounted) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('Profile photo removed'),
-              backgroundColor: Colors.green,
-            ),
-          );
-
-          // Reset the photo.
-
-          setState(() {
-            _profilePhoto = null;
-          });
-
-          // Notify parent of data change.
-
-          widget.onDataChanged();
-        }
-      }
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isUploadingPhoto = false;
-        });
-      }
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    const int notificationCount = 2;
+    // Get the name controller for display purposes.
+
+    final nameController = _controllers.keys.firstWhere(
+      (controller) => _controllers[controller] == 'name',
+    );
 
     return Container(
       constraints: const BoxConstraints(
@@ -1005,153 +340,27 @@ class _ProfileDetailsState extends State<ProfileDetails> {
             children: [
               // Title and edit button row.
 
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Text(
-                    'Profile Details',
-                    style: TextStyle(
-                      fontSize: 18,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  if (widget.showEditButton)
-                    MarkdownTooltip(
-                      message: '''
-
-                      **Edit Profile Details**
-
-                      Click to modify your personal information:
-
-                      - Name
-                      - Address
-                      - Contact information
-                      - Personal details
-
-                      Your data is securely stored in your personal pod.
-
-                      ''',
-                      child: IconButton(
-                        icon: const Icon(Icons.edit),
-                        onPressed:
-                            _isLoading || _isSaving ? null : _showEditDialog,
-                      ),
-                    ),
-                ],
+              ProfileUIComponents.buildTitleRow(
+                context,
+                widget.showEditButton,
+                _isLoading,
+                _isSaving,
+                _showEditDialog,
               ),
+
               const SizedBox(height: 12),
 
               // Avatar and Name Section
               if (_isLoading)
                 const Center(child: CircularProgressIndicator())
               else
-                Row(
-                  children: [
-                    // User avatar with lock icon.
-
-                    Stack(
-                      clipBehavior: Clip.none,
-                      children: [
-                        // Profile photo with loading indicator or initials.
-
-                        ProfilePhotoHandler.buildProfileAvatar(
-                          context: context,
-                          photo: _profilePhoto,
-                          name: _nameController.text,
-                          radius: 24,
-                          isLoading: _isLoadingPhoto || _isUploadingPhoto,
-                          onTap: _showPhotoOptionsDialog,
-                        ),
-
-                        // Security lock indicator.
-                        Positioned(
-                          bottom: -2,
-                          right: -2,
-                          child: Container(
-                            padding: const EdgeInsets.all(2),
-                            decoration: BoxDecoration(
-                              color: theme.colorScheme.tertiary,
-                              shape: BoxShape.circle,
-                            ),
-                            child: Icon(
-                              Icons.lock,
-                              color: theme.colorScheme.onTertiary,
-                              size: 16,
-                            ),
-                          ),
-                        ),
-
-                        // Edit photo indicator.
-
-                        if (!_isLoadingPhoto && !_isUploadingPhoto)
-                          Positioned(
-                            top: -2,
-                            left: -2,
-                            child: Container(
-                              padding: const EdgeInsets.all(2),
-                              decoration: BoxDecoration(
-                                color: theme.colorScheme.surface,
-                                shape: BoxShape.circle,
-                                border: Border.all(
-                                  color: theme.colorScheme.primary,
-                                  width: 1,
-                                ),
-                              ),
-                              child: Icon(
-                                Icons.edit,
-                                color: theme.colorScheme.primary,
-                                size: 12,
-                              ),
-                            ),
-                          ),
-                      ],
-                    ),
-                    const SizedBox(width: 12),
-
-                    // Display user name.
-
-                    Expanded(
-                      child: Text(
-                        _nameController.text,
-                        style: theme.textTheme.titleMedium?.copyWith(
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    ),
-
-                    // Notification bell with notification count badge.
-
-                    Stack(
-                      clipBehavior: Clip.none,
-                      children: [
-                        Icon(
-                          Icons.notifications,
-                          size: 28,
-                          color: theme.colorScheme.onSurfaceVariant,
-                        ),
-                        // Notification counter badge.
-
-                        if (notificationCount > 0)
-                          Positioned(
-                            right: -2,
-                            top: -2,
-                            child: Container(
-                              padding: const EdgeInsets.all(4),
-                              decoration: BoxDecoration(
-                                color: theme.colorScheme.error,
-                                shape: BoxShape.circle,
-                              ),
-                              child: Text(
-                                '$notificationCount',
-                                style: theme.textTheme.labelSmall?.copyWith(
-                                  color: theme.colorScheme.onError,
-                                ),
-                              ),
-                            ),
-                          ),
-                      ],
-                    ),
-                  ],
+                ProfileUIComponents.buildProfileHeader(
+                  context,
+                  _profilePhoto,
+                  nameController.text,
+                  _isLoadingPhoto,
+                  _isUploadingPhoto,
+                  () => _showPhotoOptionsDialog(),
                 ),
 
               const SizedBox(height: 16),
@@ -1161,42 +370,14 @@ class _ProfileDetailsState extends State<ProfileDetails> {
               // Personal Identification Details section.
 
               if (_isLoading)
-                ..._buildLoadingRows()
+                ...ProfileUIComponents.buildLoadingRows(context)
               else
                 Form(
                   key: _formKey,
                   child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     mainAxisSize: MainAxisSize.min,
-                    children: [
-                      _buildDataRow('Address:', _addressController.text),
-                      const SizedBox(height: 6),
-                      _buildDataRow('Phone:', _bestContactPhoneController.text),
-                      const SizedBox(height: 6),
-                      _buildDataRow(
-                        'Emergency Name:',
-                        _emergencyNameController.text,
-                      ),
-                      const SizedBox(height: 6),
-                      _buildDataRow(
-                        'Emergency Phone:',
-                        _emergencyPhoneController.text,
-                      ),
-                      const SizedBox(height: 6),
-                      _buildDataRow(
-                        'Alternative:',
-                        _alternativeContactNumberController.text,
-                      ),
-                      const SizedBox(height: 6),
-                      _buildDataRow('Email:', _emailController.text),
-                      const SizedBox(height: 6),
-                      _buildDataRow(
-                        'Date of Birth:',
-                        _dateOfBirthController.text,
-                      ),
-                      const SizedBox(height: 6),
-                      _buildDataRow('Gender:', _genderController.text),
-                    ],
+                    children: _buildDataRows(),
                   ),
                 ),
             ],
@@ -1204,102 +385,70 @@ class _ProfileDetailsState extends State<ProfileDetails> {
           // Loading/saving overlay.
 
           if (_isLoading || _isSaving)
-            Positioned.fill(
-              child: Container(
-                color: theme.cardTheme.color?.withValues(alpha: 0.7),
-                child: Center(
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      const CircularProgressIndicator(),
-                      const SizedBox(height: 8),
-                      Text(
-                        _isLoading
-                            ? 'Loading profile data...'
-                            : 'Saving profile data...',
-                      ),
-                    ],
-                  ),
-                ),
-              ),
+            ProfileUIComponents.buildLoadingOverlay(
+              context,
+              _isLoading,
+              _isSaving,
             ),
         ],
       ),
     );
   }
 
-  /// Create placeholder loading rows during data fetch.
+  /// Builds data rows for all profile fields.
 
-  List<Widget> _buildLoadingRows() {
-    return List.generate(
-      6,
-      (index) => Padding(
-        padding: const EdgeInsets.symmetric(vertical: 8.0),
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            SizedBox(
-              width: 100,
-              child: Container(
-                height: 14,
-                width: 80,
-                decoration: BoxDecoration(
-                  color: Theme.of(context)
-                      .colorScheme
-                      .onSurfaceVariant
-                      .withValues(alpha: 0.2),
-                  borderRadius: BorderRadius.circular(2),
-                ),
-              ),
-            ),
-            Expanded(
-              child: Container(
-                height: 14,
-                decoration: BoxDecoration(
-                  color: Theme.of(context)
-                      .colorScheme
-                      .onSurfaceVariant
-                      .withValues(alpha: 0.1),
-                  borderRadius: BorderRadius.circular(2),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
+  List<Widget> _buildDataRows() {
+    final fieldLabels = {
+      'address': 'Address:',
+      'bestContactPhone': 'Phone:',
+      'emergencyName': 'Emergency Name:',
+      'emergencyPhone': 'Emergency Phone:',
+      'alternativeContactNumber': 'Alternative:',
+      'email': 'Email:',
+      'dateOfBirth': 'Date of Birth:',
+      'gender': 'Gender:',
+    };
+
+    final rows = <Widget>[];
+    _controllers.forEach((controller, fieldName) {
+      final label = fieldLabels[fieldName];
+      if (label != null) {
+        rows.add(
+          ProfileUIComponents.buildDataRow(context, label, controller.text),
+        );
+        rows.add(const SizedBox(height: 6));
+      }
+    });
+
+    // Remove the last SizedBox.
+    if (rows.isNotEmpty) {
+      rows.removeLast();
+    }
+
+    return rows;
   }
 
-  /// Build a single data row with label and value.
+  /// Shows photo options dialog using ProfilePhotoManager.
 
-  Widget _buildDataRow(String label, String value) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SizedBox(
-          width: 100,
-          child: Text(
-            '$label ',
-            style: const TextStyle(
-              fontWeight: FontWeight.bold,
-              fontSize: 13,
-            ),
-          ),
-        ),
-        Expanded(
-          child: Text(
-            value.isEmpty ? '—' : value,
-            style: TextStyle(
-              color: value.isEmpty
-                  ? Theme.of(context)
-                      .colorScheme
-                      .onSurfaceVariant
-                      .withValues(alpha: 0.5)
-                  : Theme.of(context).colorScheme.onSurface,
-            ),
-          ),
-        ),
-      ],
+  Future<void> _showPhotoOptionsDialog() async {
+    final nameController = _controllers.keys.firstWhere(
+      (controller) => _controllers[controller] == 'name',
+    );
+
+    await ProfilePhotoManager.showPhotoOptionsDialog(
+      context,
+      _profilePhoto,
+      nameController.text,
+      (photo) {
+        setState(() {
+          _profilePhoto = photo;
+        });
+      },
+      widget.onDataChanged,
+      _isLoading,
+      _isSaving,
+      _isLoadingPhoto,
+      _isUploadingPhoto,
     );
   }
 }
